Update 0500-10-12-mapathon_en.md
Toc partially not working. I am not shure what the root cause is. But i noticed that the items that are not working contain at least one dot.

So i removed the dots, but could't test the outcome: the preview doesn't render the toc.

Please let me know if it worked, thanks.